Simple demo of an in order traversal
Happens on a binary tree, the first example that comes up in TAOCP 1.
More details on the algorithm used can be found here:
https://github.com/adammccartney/algorithms/commit/bcf8d736a32d9deca2385d7ddb73b0607eace45b

The program performs the traversal correctly, the current stumbling
blocks are happening when trying to write a dump function that displays
the shape of the tree somehow.